Major upgrade of Lifegraph, which can now display a variety of graphs. Evaluation screen uses the new Lifegraph system to display level, life and score graphs. Use the BEATLINE key to cycle them. For Coop, Team and Vs Cpu modes, combined scores are shown in the life graph. Evaluation screen now shows overall performance percentages in Normal mode, Peak Groove Momentum in Coop mode, and Victory/Defeat in Team and Vs Cpu mode. Started implementation of record  (high score) replaying.

+++ WindowsGame1/WindowsGame1/Drawing/RoundLine/RoundLine.cs
@@ -472,7 +472,7 @@ namespace RoundLineCode
         /// <param name="roundLines">A collection of RoundLines to draw.</param>
         /// <param name="lineRadius">The radius (half width) of the lines to draw.</param>
         /// <param name="lineColor">The colour of the lines to draw.</param>
-        public void Draw(IEnumerable<RoundLine> roundLines, int lineRadius, Color lineColor)
+        public void Draw(IEnumerable<RoundLine> roundLines, float lineRadius, Color lineColor)
         {
             Draw(roundLines,lineRadius,lineColor,0,null);
         }
